Baremetal: Add build for AVR
This commit adds a baremetal build environment for AVR MCUs.
Specifically, we add test/baremetal/platform/avr which includes an
additional Makefile -- to be specified via EXTRA_MAKEFILE -- which
compiles mlkem-native for a version of Atmega128rfr2 with 32K instead
of 16K RAM, and executes it using the simavr simulator.

16K of RAM are not enough for for mlkem-native at the moment -- we
should look into improving this -- and the Atmega256rfr2, which has
32K of RAM, is not supported by simavr. To remedy, we patch simavr
to use 32K instead of 16K RAM for Atmega128rfr2. The AVR toolchain
and patched version of simavr are provided via a new nix shell `avr`.

The point of this baremetal target is less to suggest that mlkem-native
is ready, or ready yet, for application on MCUs. Instead, the primary
purpose is to build in a 16-bit C environment and to force-identify any
places in the library where we rely on `int`/`unsigned` being at least
32-bit wide, and adjusting the code accordingly.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>