FIPS202: Use `unsigned` for hash rate
This commit changes the signatures of the internal functions
`absorb_once`, `absorb_once_x4`, `squeezeblocks`, `squeezeblocks_x4`
to provide the hashing rate `r` as an `unsigned` rather than a
uint32. Even if `unsigned` is 16-bit, it is enough to hold the
hashing rate, while on 16-bit systems, using `uint32_t` leads to
integer conversion warnings when calling into the KeccakF1600 API
which works with `unsigned` throughout.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>