Adjust for 16-bit C implementations
This commit ensures that mlkem-native is functionally correct
on C implementations where `int` and `unsigned` are 16-bit wide.

On such implementations, integer promotion is to a 16-bit domain,
which is not enough for all computations performed in mlkem-native.
Specifically, the scalar compression functions rely on interim 32-bit
arithmetic, as does Barret reduction. This is fixed by explicitly
promoting operands to (u)int32_t.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>