AVR: Add support for command line arguments
This commit extends the exec wrapper for AVR emulation with support
for command line arguments. An EEPROM image is built from the command
line and the AVR runtime amended to copy the EEPROM into RAM and setting
up the argc and argv arguments for main (care has to be taken to provide
a contiguous EEPROM hex file -- simavr does not handle disconnected
regions correctly).

This allows to run the ACVP tests on AVR. However, since the amount of
space needed for the command line argument (in hex format) is significant
(>8K), we need to patch simavr another time to increase the EEPROM size
for atmega128rfr2.

We also split the number of test iterations in KAT and func tests into
separate macros, from NTESTS in both to NTESTS_KAT and NTESTS_FUNC,
respectively. NTESTS_FUNC can then be overwritten in the CFLAGS without
affecting the correctness of the KAT tests, which rely on a fixed number
of iterations because its entire output is hashed and compared to META.yml.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>