CI/CBMC: Don't repeat level-independent proofs
This commit modifies the CBMC CI to not check proofs of level-independent
functionality three times for MLKEM-{512,768,1024}.

Rather than trying to find out exactly which functions are level-independent
vs. level-agnostic, we only list a few files which we are sure only to include
level-agnostic functionality, and exclude the functions whose contracts is
declared in those files. We then skip the proofs for those functions in the
CBMC CI runs for MLKEM-768 and MLKEM-1024. By default, however, `tests cbmc`
continues to run all proofs.

We also drop functionality tests from the CBMC CI runs.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>