CI ec2_functests: Do not use custom AMIs
Currently ec2_functests are using custom AMIs making it harder to port it to
mldsa-native. I suspect that the only good reason for using a custom AMI was
to increase the volume size from the default 8 GB, but we may as well do that
using the ec2_volume_size argument.
This commit switches away from custom AMIs.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>