Speed up `make`
A plain `make clean` would previously take >3s, primarily because
of an unnecessary use of dozens of shell invocations that could
be replaced by a mere `subst` inside `make`.

Also, skip the bulk of the makefiles when cleaning, and align
the output format of `make clean` with that of other targets.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>