Make allocation of large structures configurable
This commit introduces the configuration option

  MLK_CONFIG_CUSTOM_ALLOC_FREE

which allows users to provide custom macros

  MLK_CUSTOM_ALLOC(v, T, N)
  MLK_CUSTOM_FREE(v, T, N)

that should be used in place of the default stack-allocation
to allocate/free large internal structures. Those macros are
then wrapped into MLK_ALLOC/MLK_FREE and used in the source.
Importantly, MLK_FREE adds zeroization ahead of calling
MLK_CUSTOM_FREE, so the latter need not take care of this.

The macros are put to use in kem.c and indcpa.c, but not yet
further down the call-stack.

The option is documented as experimental and (hence) unstable
so we have freedom to adjust this ahead of v2.

A custom configuration is added which implements the macros
based on C11's `aligned_alloc`. We enable the leak sanitizer
in the tests for this configuration to catch any missing calls
to MLK_FREE.

Since allocation can fail, additional error paths are introduced
to the functions using MLK_ALLOC. This uniformly follows the pattern
of all pointers being initialized to NULL before allocation, and a
cleanup section calling MLK_FREE on them before returning the result.

To distinguish between out-of-memory failure and other kinds of
functional failures, we introduce named error codes MLK_ERR_FAIL
and MLK_ERR_OUT_OF_MEMORY. This will likely need further refinement,
but this can only be done in v2.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>