Bundle multiple allocations into workspaces
This commit introduces 'workspace structures', which are
are function-local structures wrapping all objects and buffers
utilized by the function into a single struct.

The benefit of the workspace struct is a simpler allocation
boilerplate: Instead of allocating multiple structures and
separately handling their success/failure, we now only need
to allocate a single object per function.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>