Fix pointer cleanup on dynamic allocation failure
When using a custom alloc/free implementation, allocation may fail,
in which case the cleanup section needs to zeroize and free those
pointers there were successfully allocated, and skip those for which
allocation failed. This skipping of failed allocations is not
implemented correctly and would lead to a null pointer dereference.

This commit fixes this.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>