CI: Add native ppc64le (POWER10) CI
Courtesy of Basil Hess/IBM, we know have a self-hosted POWER10 runner.
This commit adds functional tests to CI.
I had attempted to add it to the existign "kat_tests", but the sanitizer tests
don't work properly on that platform.
Note that opt tests are already enabled so that
https://github.com/pq-code-package/mlkem-native/pull/1193 can simply be
rebased on top of this.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>