x86_64: Autogenerate the entire constant array qdata for NTT/INTT
The constant array required for the x86_64 consists of multiple parts:
first some index constants, then the twiddle factors for the NTT/INTT, and
then the twiddle factors for the mulcache computation.
Thus far, the autogeneration script would only generate the twiddle factors
into two separate .i files that are then included into consts.c, while the
index constants are hardcoded in the consts.c

However, for HOL-Light we require to generate the entire constant array.
Having two diffent functions generating the constants is a risk that they may
eventually diverge.
This commit fixes this by autogenerating the entire qdata in consts.c
(the .i files is no longer needed). This way we can use the same
generation function for both consts.c and the HOL-Light proofs.

While at it we also improve documentation of the index constants and also
auto-generate the offsets in consts.h

- Resolves https://github.com/pq-code-package/mlkem-native/issues/1428
- Resolves https://github.com/pq-code-package/mlkem-native/issues/1380

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>