test_alloc: Align bump buffer
test/test_alloc.c uses a custom bump allocator. While all allocation sizes
are aligned to multiples of 32 bytes, the allocation buffer itself can be
unaligned. This can crash the test on systems with alignment requirement,
such as x86+AVX2.

This is not an issue in the actual implementation, which uses appropriate
alignment constraints: See the default implementation of MLK_ALLOC in
src/common.h.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>