tests: Add alloc tests to tests script and functest/multi-functest Action
In https://github.com/pq-code-package/mlkem-native/pull/1424 it was missed
to add the alloc tests to the scripts/tests script and consequently not
all tests in CI were including the alloc tests.

This commit adds the alloc tests to the script and also extends the functest
and multi-functest with a alloc input enabling/disabling the alloc test.

This should catch bugs like https://github.com/pq-code-package/mlkem-native/issues/1433.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>